+++ snowddl/resolver/__init__.py
@@ -24,6 +24,7 @@ from .foreign_key import ForeignKeyResolver
 from .function import FunctionResolver
 from .hybrid_table import HybridTableResolver
 from .iceberg_table import IcebergTableResolver
+from .join_policy import JoinPolicyResolver
 from .masking_policy import MaskingPolicyResolver
 from .materialized_view import MaterializedViewResolver
 from .network_policy import NetworkPolicyResolver
@@ -110,6 +111,7 @@ default_resolve_sequence = [
     # --
     AggregationPolicyResolver,
     AuthenticationPolicyResolver,
+    JoinPolicyResolver,
     MaskingPolicyResolver,
     NetworkPolicyResolver,
     ProjectionPolicyResolver,
@@ -180,6 +182,7 @@ singledb_resolve_sequence = [
     # --
     DatabaseRoleResolver,
     AggregationPolicyResolver,
+    JoinPolicyResolver,
     MaskingPolicyResolver,
     ProjectionPolicyResolver,
     RowAccessPolicyResolver,